+++ Assets/Scripts/Entity/StatusEffect/Buff/DEFUP.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NewDEFUP", menuName = "StatusEffect/DEFUP")]
 public class DEFUP : StatusEffect
 {
-    public int PercentAmount;
+    public float PercentAmount;
 
     public override void OnApply(CharacterInBattle target)
     {
